Don't process weight refunds
